TF-7 Add ability to customise favicon
Add the ability to add custom favicons to blob storage which are
picked up on startup and replace the (new) default favicon.

The favicons are stored in memory for speed and cached on the
client side for an hour.

Also fixed an issue where the bottom underline of a thread with no
replies wasn't showing.